requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
PyMuPDF
sentence-transformers
faiss-cpu
